+++ NEWS.md
@@ -1,3 +1,13 @@
+# spant 1.7.0
+* Added write_mrs function which guesses the output format from the file
+extension or can be specified as an argument. write_mrs_XXX functions have been
+depreciated.
+* read_mrs function now tries to guess the format from the file extension.
+* Added json sidecar to NIFTI MRS export function.
+* Added the option to read MRS data from a NIFTI file and json sidecar using the
+read_mrs function.
+* Changed default crop_spec region to between 4.0 and 0.2 ppm.
+
 # spant 1.6.0
 * Bug fix for GE P file reader.
 * Added downsample_mrs function.